Bitwise op

+++ inference.py
@@ -225,7 +225,7 @@ def main():
                                       apply_dilatation=False)
                 cv2.imwrite("./results/mouth_{}.png".format(idx), 255 * np.uint8(bottom_mask))
                 bottom_mask = bottom_mask > 0
-                cv2.imwrite("./results/mouth_2_{}.png".format(idx), bottom_mask)
+                cv2.imwrite("./results/mouth_2_{}.png".format(idx), np.uint8(bottom_mask))
                 #nose = lm[idx][27:35+1]
                 #nose_mask = np.zeros_like(ff)
                 #element = np.ones((3,3), dtype=np.uint8)
